requirements.txt
sentence-transformers
datasets
transformers
wandb
openai
numpy
python-dotenv
huggingface-hub
torch
